move result UI to production
