+++ QuizApp/QuizApp/View/CorrectAnswerCell.swift
@@ -1,0 +1,14 @@
+//
+/*
+ *		Created by 游宗諭 in 2021/8/22
+ *
+ *		Using Swift 5.0
+ *
+ *		Running on macOS 12.0
+ */
+
+import UIKit
+class CorrectAnswerCell: UITableViewCell {
+    let questionLabel = UILabel()
+    let answerLabel = UILabel()
+}
