+++ QuizApp/QuizApp/View/PresentableAnswer.swift
@@ -1,0 +1,6 @@
+struct PresentableAnswer {
+    let question: String
+    let answer: String
+    let wrongAnswer: String?
+    var isCorrect: Bool { wrongAnswer == nil }
+}
