+++ QuizApp/QuizApp/View/WrongAnswerCell.swift
@@ -1,0 +1,16 @@
+//
+/*
+ *		Created by 游宗諭 in 2021/8/22
+ *
+ *		Using Swift 5.0
+ *
+ *		Running on macOS 12.0
+ */
+
+import UIKit
+
+class WrongAnswerCell: UITableViewCell {
+    let questionLabel = UILabel()
+    let correctAnswerLabel = UILabel()
+    let wrongAnswerLabel = UILabel()
+}
